+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
-All notable changes to the "unique-window-colors" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+All notable changes to the "git-repo-window-colors" extension will be documented in this file.
 
-## [Unreleased]
+## [1.1.0]
+
 - Initial release
+
+## [1.2.0]
+
+- You can apply branch coloring without requiring a repo rule match first.